The bug technically lies in route_table_policy.cc but I preferred a more generic fix.
If a route is modified but rejected, the caller should obtain the filtered
message and delete it. This does not occur in route_table_policy.cc. A scenario
depicting this is actually present in policy/test/policy2.src.
If BGPVarRW was used, there would have been a memory leak.

As a fix, the BGPVarRW will check if the caller obtained the filtered message.
If it didn't, BGPVarRW will delete the filtered message upon its destruction.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4711 26435e26-4b25-483a-942d-477236992e84