Refactor the implementation of IPv6::make_prefix() by moving the static array v6prefix inside IPv6::make_prefix().
The refactoring is needed because in case of OpenBSD with egcc-3.2.2
compiler the static v6prefix[] array may loose state (probably
because of a bug with the compiler)

For example, the first time IPv6::make_prefix() is called,
it correctly initializes the v6prefix vector with the appropriate
values. However, the second time IPv6::make_prefix() is called,
the values inside the v6prefix[] vector are lost; this compiler bug
cannot be triggered by test_ipv6 itself, but when running, say,
pim/test_pim, or some of the rip/test_* test programs.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4732 26435e26-4b25-483a-942d-477236992e84