Improve RIB regression testing in preparation for blackhole routes.
Expand the RIB's verify_route() syntax to determine if a given destination
has no route, has an IP family route, or has a route whose next-hop is
a DiscardNextHop (i.e. is a blackhole route).

As a pleasant side-effect, we eliminate the special-casing of the vifname
field whereby we would check for a non-existent interface called 'discard'.
This was also somewhat misleading as the condition being checked for was
in fact the absence of a route for the given destination.

Apply the following substitutions to the 'commands' script used to
drive the RIB regression tests to make use of the new syntax:

1)
verify
becomes
verify ip

2)
^route verify ip \([0-9.]*\) discard \(.*\)
becomes
route verify miss \1 lo0 0.0.0.0 0

Apply the following regexps to test_register.cc:

1)
test_route_range("\([0-9.]*\)", \(.*\)
becomes
test_route_range("\1", "ip", \2

2)
test_route_range("\([0-9.]*\)", "ip", "discard", \(.*\)
becomes
test_route_range("\1", "miss", "lo0", \2


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4850 26435e26-4b25-483a-942d-477236992e84