Stub out the DiscardNextHop code which was introduced in rev 1.40 to RIB<A>::verify_route().
This is only good for routes which forward traffic
via unnumbered interfaces, which we don't currently support;
to avoid refactoring much of the RIB code, we force the
notion of discard routes to be implemented using numbered
Vifs.

Therefore, discard routes do not require to be treated as
a special case; they may be treated like any other
'connected' route, i.e. one which directly references a
Vif and does not require next-hop resolution.

Add some explanatory comments and stubbed-out code for
the future DiscardNextHop case (likely to be renamed
to UnnumberedNextHop or InterfaceNextHop) in
RIB<A>::lookup_route().

Discussed with:	pavlin


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4862 26435e26-4b25-483a-942d-477236992e84