Remove an unnecessary cached iterator. i2 is assigned to the current value of i, then i is incremented. The element pointed to by the i2 iterator is then erased.
Also: Delete a pointed-to element before erasing its
pointer which is held in a sequenced container, as
the iterator will then be invalid.

This code is marked unreachable, so this fix does
nothing, but will stop us from tripping over this
in future bug sweeps.

Glanced at by:  mark, pavlin


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4881 26435e26-4b25-483a-942d-477236992e84