Add two new methods to interface fea_fib_client/0.1:     resolve_route4, resolve_route6
This is a notification from the FEA to indicate
that there was a route miss or a request for
next-hop resolution from the forwarding plane.

It exists primarily to support on-demand and
ad-hoc routing protocols implemented using
the XORP framework.

Add code to present clients of this interface
(currently only fib2mrib) to ignore this event.

Platform support for generating this event
will be forthcoming in future commits.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4910 26435e26-4b25-483a-942d-477236992e84