Final round of changes to support on-demand and ad-hoc routing protocols.
Add the appropriate XRL glue to pass RESOLVES to client
processes of the FEA FIB.

Add the appropriate plumbing to do this for BSD rtsock.
Specifics are as follows:

Add and make use of a new integer typedef, FtiFibMsgSet.

Flip the sense of the current filtering to be inclusive
rather than exclusive, so we can deal with those situations
where routing table observers want to hear about a wider
variety of events.

Callers within the FEA can ask for UPDATES (adds, deletes,
changes), GETS (the 'get' message), and RESOLVES (route
misses and resolves).

When we get routes back in the FtiConfigTableObserver,
look for and flag unresolveds in the copies we make before
we hand them off via XRLs, so that the XRL glue will know
what to do with them.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4929 26435e26-4b25-483a-942d-477236992e84