Do not build the XORP test programs by default.
Instead, add a dependency for them to the top-level
check target using the automake 'check' prefix,
such that they are built at 'gmake check' time.

This reduces the pre-install disk space requirements
for a standard XORP build on FreeBSD 4.10-RELEASE
from 1.91GB to 1.28GB.

This does not apply to the test XRL targets or their
associated libraries.

Regen the Makefile.in files.

Approved by: pavlin


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4974 26435e26-4b25-483a-942d-477236992e84