Add code that keeps an audit trail of the last 1000 events to pass through a dump table.
There's still a bug somewhere in BGP that causes occasional aborts
during the process of dumping to new peers, as found in crash17 on
xorp12.  There's not enough data in the crash17 core dump to diagnose
the history leading up to this, so this audit trail is an attempt to
keep enough data.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@4986 26435e26-4b25-483a-942d-477236992e84