Add a mechanism to call an external program to generate the Click configuration whenever the network information changes.
Note: this should be an intermediate solution, because
design-wise the FEA should not call external programs on-the-fly.
A more appropriate design solution is that the FEA should always
receive the generated configuration (e.g., via XRL). For example,
when the interface configuration is changed via the xorpsh,
the xorpsh itself may call the external program to regenerate
the new Click configuration, and then send that configuration
by XRL to the FEA. Currently, we cannot use this solution, because
the xorpsh/rtrmgr configuration template semantics do not provide for
externalgenerator programs whose output can be used as an XRL argument.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5004 26435e26-4b25-483a-942d-477236992e84