Change the way xorpsh learns the initial router config from the rtrmgr.  Previously it requested it.  Now it's proactively given it via the rtrmgr_client interface as soon as it's finished authentication.  This avoids a possible race where a config change could occur before xorpsh knew about the config, which would have caused a crash.
Add code to proactively report the current running modules and their
status as soon as xorpsh finished authenticating, and to update this
information as it changes.  This information will (eventually) be used
to ensure that only relevant operational mode commands are available
through the CLI, not ones for protocols we're not currently running.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5070 26435e26-4b25-483a-942d-477236992e84