Check the return value of close(2) after writing to a Click kernel handler.
If the close() system call returned an error, this is
an indication that the write to the Click handler failed.

Note that typically we should have caught any errors by
reading the "<click>/errors" handler, so the check for
the return value of close() is the last line of defense.

Obtained from:	Eddie Kohler


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5076 26435e26-4b25-483a-942d-477236992e84