Major refactoring of class hierarchy in rtrmgr and xorpsh.
Previously xorpsh needed to be linked with lots of code from rtrmgr
that was not used, just to satisfy compilation dependencies, even
though the code was not used.  The code was getting too hard to work
with when anything needed to be added to xorpsh.  This refactoring
removes almost all of the unused code, by introducting more
specialized subclasses of many of the classes for use in rtrmgr
(prefixed Master-).  The less specialized versions continue to be used
in xorpsh.

This commit touches a lot of code.  It has been tested, but not
desperately thoroughly.  Next step is to modify the test code which is
currently not enabled to use the refactored classes.  Currently the
tinderbox doesn't test anything in rtrmgr, which is not good.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5100 26435e26-4b25-483a-942d-477236992e84