XORP code uses typedefs (e.g. uint32_t, int32_t) rather than using the base types, because the 'C' language allows a compiler to use a natural size for base type. The XORP code is internally consistent in this usage, one problem arises with format strings in printf style functions.
In order to print a size_t or uint32_t with "%u" it is necessary to
cast to an unsigned int. On Mac OS X a size_t is not an unsigned
int. On windows uint32_t is not an unsigned int.

In order to print a int32_t with a "%d" it is necessary to cast to
a signed int. On windows int32_t is not a signed int.

Two convenience macros to help with the casting.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5115 26435e26-4b25-483a-942d-477236992e84