Linux coredump (temporary) fix:
All MasterConfigTreeNode local state has been moved
temporary to the base ConfigTreeNode class.
This is a temporary solution of a memory allocation problem:
in number of places, ConfigTreeNode objects are allocated,
but later they are casted (using C-style cast) to the larger
MasterConfigTreeNode, and the extra state (outside of the allocated
memory) is modified.

Approved by:	Atanu


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5121 26435e26-4b25-483a-942d-477236992e84