Add support for removing as well as adding an optional parameter. Now a setting such as "ipv4-multicast" is a variable that can be enabled or disabled.
Now that it is possible to remove a parameter it is possible to
configure BGP to not send IPv4 Unicast traffic. Previously we would
send to a peer whatever it had requested a AFI/SAFI now we send what
we and the peer have agreed on.

The changes above have uncovered a number of bugs that have also been
fixed in this commit.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5141 26435e26-4b25-483a-942d-477236992e84