A little more refactoring in the ModuleManager.
Fix so that xorpsh module manager gets updated properly with the
module state from the rtrmgr.

Use module state in xorpsh to determine whether or not to provide the
operational mode commands for each module to the user.  Only commands
for active modules will be shown.  This should fix bug 26.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5173 26435e26-4b25-483a-942d-477236992e84