BGP uses the endpoint addresses and ports used by a peering to define a peering. The protocol actually uses fixed ports variable ports are used for debugging; allowing more than one session between a pair of hosts.
A class Iptuple takes the addresses and ports and defines a
peering. An address is in textual form allowing IPv4 and IPv6 notation
as well as hostnames. If a lookup fails then Iptuple will throw an
exception. Iptuple can also throw an exception if that address
families of the two addresses don't match.

The exceptions thrown by Iptuple are now been caught, a bad address
will now generate an error, not cause BGP to exit.

This is part of the fix for bug:
<http://www.xorp.org/bugzilla/show_bug.cgi?id=105>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5206 26435e26-4b25-483a-942d-477236992e84