Implement the following rtrmgr template command:
%deprecated: "String with reason for template node deprecation"

This command can be used to deprecate a template tree node
(and the subtree below it).

If the XORP startup configuration contains a statement that uses a
deprecated node in the template, the rtrmgr prints an error with
the string with the reason, and exits.
If, however, somehow the xorpsh sends to the rtrmgr configuration
that contains deprecated statement, the rtrmgr returns an error,
and the error string returned to the xorpsh will contain
the %deprecated string.

Obviously, the %deprecated string can contain anything; e.g.,
instructions to the user what configuration to use as a
substitution, etc.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5244 26435e26-4b25-483a-942d-477236992e84