Improve the support for using the doc-related macros when compiling in a directory that is different from the source directory.
Note that this is not the final solution. E.g., if the Latex file
includes a file with a figure, unlike C/C++ compilers we cannot
use the equivalent -I <dir> to specify the directory to search for
the included file.
A hackish work-around would be to create temporary symbolic links
to the directories with the figures, bib files, etc.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5247 26435e26-4b25-483a-942d-477236992e84