Bug fix: when deleting a routing entry in Click, allow the delete to proceed even if we don't have the next-hop info (which in case of Click translates into an outgoing port number of the lookup element).
Currently, we don't have the next-hop info when deleting a route,
simply because the XRL from the RIB doesn't carry it.

In any case, in Click if we use the "remove ADDR/MASK" command
to the LinearIPLookup element, then we are going to remove
all routes with the same ADDR and MASK. However, in case of
XORP+Click we never attempt to install more than route with the
same ADDR and MASK, therefore the "remove ADDR/MASK" format
is acceptable.

Bug found by:	Rafael Paoliello Guimaraes <rafael.guimaraes@ac.upc.edu>
Approved by:	Atanu


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5278 26435e26-4b25-483a-942d-477236992e84