Don't hard-wire a "static" table inside RIB.
The reason for this hard-wiring is because if RIP is configured to
export the static routes, it cannot do it if the "static" origin table
is not there already.
The work-around solution is to modify the rtrmgr templates and to
make RIP depend on "static_routes".

This fixes the remaining issues in Bugzilla entry #69

Approved by:	Atanu


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5386 26435e26-4b25-483a-942d-477236992e84