Fix two problems related to next hop resolutions changing and add a regression test that used to trigger the problem.
When the decision process finds possible routes it checks the
resolvability of the route by querying the next hop resolver. Routes
themselves carry resolvability information. It is possible that a
route may become resolvable but is still marked as unresolved. In the
decision process update the resolvability state of all candidate
routes.

In the RIB-OUT it is now legal to get an add followed by a delete.
The happens when a route with the same nexthop has been introduced on
two peers.

The RIB-IN sends a delete followed by an add. If this is
the better route then the other peer will send an add then
a delete.

RIB-IN				RIB-OUT
Peer A -> delete<A>
				add<B>
Peer A -> add<A>
				delete<B>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5470 26435e26-4b25-483a-942d-477236992e84