Refactor the mechanism for checking whether the nexthop is directly connected. The refactoring fixes a bug when we are adding a "connected" route and there are already routing entries in the RIB (e.g., if a new IP address is added to a network interface).
Previously, we would lookup the nexthop address in the final table,
and then we would use the corresponding vif to check whether the
nexthop address is directly connected to that vif.
However, this doesn't work if we are adding a "connected" route
(which may use the local IP address as the next-hop router),
and there is already, say, 0/0 routing entry. The lookup will
return the vif toward the default router, but then the next-hop
address of the "connected" route won't match as being
directly-connected to that vif.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5485 26435e26-4b25-483a-942d-477236992e84