Bug fix: declare a variable of type uint32_t instead of size_t.
When unpacking an XrlList argument inside XrlAtom::unpack_list(),
the local variable named "nelem" that contains the number of elements
inside an XRL must be defined as uint32_t instead of size_t,
because later the code uses sizeof(nelem) to calculate the number
of consumed bytes. On 64-bit machine, size_t is different from uint32_t,
hence the number of consumed bytes is incorrect.

Bug found on OpenBSD + amd64.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5536 26435e26-4b25-483a-942d-477236992e84