Rewrite of BGP output branch flow control.
The old code pushed changes from fanout to ribout, unless told not to.
This meant no one place was responsible for clocking data through -
sometimes it was pushed and sometimes it was effectively pulled.

The new code always queues in fanout.  Then changes are pulled from
the ribout using get_next_message().  If the ribout was idle, then
there's a new wakeup() call that fanout sends downstream to wake the
ribout up.  This all makes it much clearer who is in control, so the
code (especially in DumpTable) is much easier to understand.

All tinderbox checks appear to pass, but not yet tested in the wild.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5588 26435e26-4b25-483a-942d-477236992e84