If there is a checksum error for a PIM Register packet, and if the PIM Register packet was truncated by the kernel (e.g., in some *BSD systems), then ignore the checksum error.
The problem comes from the fact that currently the *BSD kernel
(for IPv4 and IPv6) truncates the PIM Register messages to contain
only the outer IP header, the PIM + PIM Register header, and the
inner IP header.
In the future the *BSD kernels should be modified to stop truncating
the PIM Register messages, because the gain in reduced processing
overhead is questionable. In the time being we need to deal with
such packets.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5635 26435e26-4b25-483a-942d-477236992e84