Bug fix: add code to do the right thing if we try to add an interface-specific route. Previously that code was missing, so the interface-specific routes were going into a semi-blackhole.
Tested on FreeBSD-4.9, NetBSD-1.6.1-current, OpenBSD-3.6-current
and MacOS X 10.3.
[This is routing sockets specific code so it doesn't apply to Linux]

Note that in general on {Free,Net,Open}BSD we can add a route that
has the next-hop interface AND the next-hop router set at the same
time, and the new code supports that. On MacOS X the kernel allows
us to do this only if there is already a forwarding entry for the
next-hop router address.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5706 26435e26-4b25-483a-942d-477236992e84