If the MFEA fails to transmit a protocol message and returns COMMAND_FAILED XRL error, then just print an error and continue instead of calling XLOG_FATAL().
For example, the transmission by the MFEA may fail if there is no
buffer space or if an unicast destination is not reachable.

Furthermore, all protocol messages are soft-state (i.e., they are
retransmitted periodically by the protocol),
hence we don't retransmit them here if there was an error.

This fixes a coredump problem if a neighbor router is very flaky,
or if the router is running with high capacity and is very low on
buffer space.

Bug found by:	Achmad Husni Thamrin <ahthamrin AT gmail.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5798 26435e26-4b25-483a-942d-477236992e84