Sending routes from BGP to the RIB is achieved using XRLs.
BGP allows a window of XRLs to be inflight at any time. Yesterday the
window was increased from 1 to 100. The XRL layer can refuse to accept
an XRL for transmission if for example its internal buffers are
full. If there are no route add/delete XRLs in flight and the XRL
layer refuses to accept an XRL BGP considers this to be a fatal
error. The assumption is if there are no XRLs in flight how can
the XRL layer be refusing an XRL.

This assertion was seen on xorp8 today, add some XLOG_WARNING messages
to give a hint as to why the XRLs are being refused.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5883 26435e26-4b25-483a-942d-477236992e84