Bug fix: If MRIB entries are deleted inside PIM, some of the PimMre entries may still refer to the deleted MRIB entries. This may happen during the execution of the tasks that are suppose to update the pointers to the corresponding Mrib entries (e.g., PimMre::mrib_rp() and mrib_s()).
The solution is that on startup we change the MribTable behavior to
keep all removed Mrib entries on a list instead of just deleting them.
At the end of each add/remove Mrib entries transaction,
and after we have scheduled all tasks to take care of the PimMre
updates because of the MRIB changes (i.e., the
PimMrt::add_task_mrib_changed() tasks), we add a (new) task
to delete all Mrib entries that were just removed from the MribTable.
The assumption is that by the time all actions by the
add_task_mrib_changed() task are executed, there are no more pointers
to the removed Mrib entries, and then we can safely delete them.

Bug found by:	"edrt" <edrt AT citiz.net>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5934 26435e26-4b25-483a-942d-477236992e84