Fix an AssertCancel-related issue that is in the PIM-SM spec:
According to the spec, if we receive AssertCancel on the RPF interface,
then its metric will be compared against my_assert_metric().
However, given that CouldAssert for the RPF interface is false,
my_assert_metric() will return infinite_assert_metric(), therefore
AssertCancel will fail to perform its duty.
This is fixed in the implementation by testing first if the received
Assert message contains AssertCancel metric, and if yes, the local
router is automatically declared the winner.

Note: This fix is based on the following email to the PIM Working Group
mailing list:
  Date: Fri, 9 Jul 2004 11:44:41 -0700 (PDT)
  From: Venugopal Hemige <vhemige AT yahoo.com>
  Subject: [pim] Hello HoldTime and Assert questions
  http://www1.ietf.org/mail-archive/web/pim/current/msg00206.html


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6089 26435e26-4b25-483a-942d-477236992e84