The heap code allows a set of objects to be held as a priority queue. The timer code uses the heap code. The heap code itself holds little chunks of data (heap_entry) which it holds in a contiguous block of memory. The heap_entry holds a key and a pointer to the object. When an entry is added it is placed at the end of the heap and is bubbled into position. The most recent entry is at the head of the list. If a an entry is to be removed from the heap then it would necessary to search the whole heap for it, but wait, the heap can be configured to hold an index into the heap in the object itself cool.
Both on the MIPS and Sparc the casts required to update the object
itself have been causing problems. They require casting a void * to a
char * in order to perform the arithmetic and then casting to an int *
to do the update. Pavlin points out that an extra void * cast between
the char * and int * solves the problem.

Introduce a HeapBase, objects that are to be stored on the heap
can inherit from it if they require removal from arbitary
positions.

Approved by: Pavlin


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6232 26435e26-4b25-483a-942d-477236992e84