* The neighbours can now accept LSA for transmission (queue_lsa), these LSAs are queued until push_lsas() is called.
* The Link State Request List in now a list of Ls_request not
LSa_header.

* A new method tear_down_state that is called when the state of a
neighbour drops. Stop the retransmission timer and various other tidy
activities. Most critically the LSA retransmission list is cleared, if
it wasn't it would appear that we were still awaiting ACKs and the
LSAs involved would never be candidates to enter the routing table.

* A change_state method has been added to change states, this allows
the tear_down_state method to be called if necessary.

* A new method "is_designated_router_or_backup_designated_router".

* start_rxmt_timer takes a new argument immediate. If set to true as
well as scheduling a method to be called it will call it immediately.

* A new method transmitter, in some sense the most important change
this method retransmits everything on the _lsa_request_list and
_lsa_rxmt lists.

* Two new packet sending routines:
	send_link_state_request_packet
	send_link_state_update_packet


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6302 26435e26-4b25-483a-942d-477236992e84