Implement a second mechanism for running a command inside the rtrmgr templates. Previously, "xrl" was the only supported mechanism.
The new mechanism allows us to run an external program to perform
an action. Below are three examples that illustrate its syntax:

    %create:    program "/bin/sleep 10";

    %create:    program "/bin/echo -n '$(fea.FOO)' >> /tmp/foo";

    %create:    program "$(fea.foo-program) -> stdout=$(fea.FOO_STDOUT)&stderr=$(fea.FOO_STDERR)";

In the first example the program to execute is "/bin/sleep 10".
In the second example, the "/bin/echo" program will append the value of
variable "$(rtrmgr.FOO)" as defined in the rtrmgr internal state
to file "/tmp/foo".
In the third example, the program to execute is the value of
the rtrmgr variable "$(fea.foo-program)". Its stdout and stderr output
will be written to the internal rtrmgr variables "$(fea.FOO_STDOUT)"
and "$(fea.FOO_STDERR)" respectively. Note that currently
"$(fea.FOO_STDOUT)" and "$(fea.FOO_STDOUT)" must be defined as
template variables/nodes without a type, and they are hidden to xorpsh.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6337 26435e26-4b25-483a-942d-477236992e84