Modify TemplateTree::find_node_by_type() to relax a bit the template tree node type checking.
Now if the type check failed, but if there is a matching
template node type of type NODE_TEXT, then we accept this node.

The upside of this is that we can use a single template
node like "foo @: txt" that can be used with, say,
IPv4 or IPv6 addresses, a host name, or any other text string.
The downside is that we lose the strict type checking
when our template tree contains such nodes.

Requested by:	Atanu


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6377 26435e26-4b25-483a-942d-477236992e84