The area router provides a database interface to the LSA database. In the original design it was envisaged that any code wanting to traverse the database would use the interface. Hence even invalid LSAs were returned, it was considered that it was the responsibility of the caller to do any filtering.
It looks like the only user of the database interface is the builder
of the database description packets. In which case it is worth doing
filtering in the database code itself. A problem was also noticed in
the OSPFv3 regression tests. At startup a self originating router LSA
is created. In the OSPFv3 case no links are added to the Router LSA
until the database becomes Full. Unfortunately this incomplete Router
LSA was being returned to the builder of the data description
packets. A new method has therefore been added to the LSAs
"available", which is true if a wire version of the LSA is
available. The area router database also checks with this method
before returning an LSA.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6399 26435e26-4b25-483a-942d-477236992e84