Bug fix when adding a segment to a configuration node whose template has no children.
Previously, the code assumed that the added segment is a terminal
value for a leaf node, but this sometimes allows to have garbage in
the configuration if we have a multi-value template node with
no children.
Now we explicitly check that this is a leaf node before assigning
a value to it; if not a leaf node, then this is a syntax error.

Bug found by:	Andrea Bittau


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6412 26435e26-4b25-483a-942d-477236992e84