Add support for retrieving the config operator from a node via a template file variable.
The normal template file syntax for a node's value is $(@).  The
additional syntax for a node's operator is $(<>).  Operators don't
make much sense on anything other than a terminal node, so don't
request the operator from interior nodes.

This isn't fabulously well tested, as I've no XRLs that need operators
on which to test.  I'm assuming Andrea will test this more completely.
Let me know if you have any problems with this.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6415 26435e26-4b25-483a-942d-477236992e84