Add template file support for extracting the node number of a configuration node.  The template file syntax is "$(#)"
Note that node numbers are not currently guaranteed to have any
meaningful values.  However if you load the config from a config file
(as opposed to enter it in xorpsh), the default node numbers will be
vaguely sensible.  Any nodes entered from xorpsh current get a node
number of 0.

A proper node number allocation algorithm is next on the to-do list.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6423 26435e26-4b25-483a-942d-477236992e84