attempt to fix compilation on: gcc version 4.1.0 20050618 (experimental)
WARNING: may [most likely will!!!] break compilation on other compilers!

Please inform me exactly on which compilers the dispatcher doesn't work.

This portion of code is actually the heart of the policy backend.  It is subtle,
but I would like to keep the code the way it is.  Trying to re-write it may
make it loose some power it currently has.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6443 26435e26-4b25-483a-942d-477236992e84