Fix the ~DummyFinder() destructor so now it deletes properly the _messenger.
Previously, the problem was that if we delete _messenger, its
~FinderTcpMessenger destructor tries to use a pointer to its manager()
to announce the messenger's death event.
However, at the time we call the ~DummyFinder destructor the _messenger's
manager() pointer may be invalid because the manager() may have been
deleted.

Hence, the solution is to unhook the _messenger's manager() before
deleting the _messenger itself inside the ~DummyFinder().

Note that the above problem applies only for this particular test
which defines internally the DummyFinder, and the libxipc library
itself should be fine.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6467 26435e26-4b25-483a-942d-477236992e84