Modify the ordering for sending various info by the rtrmgr to its clients (i.e., xorpsh). Now the rtrmgr sends first the module status info, and then the configuration.
Previously, the rtrmgr was sending first the configuration, and then
the module status.
However, this created a race condition in the xorpsh during startup.
Xorpsh would receive first the configuration, which would trigger its
state to become "ready" for accepting commands from the user.
Soon after that it would receive the status about which modules are
running, but in the mean time the xorpsh operational commands for those
modules weren't available.

If xorpsh is running in interactive mode, the very short delay during
startup is practically unnoticeable, but it becomes a problem if
xorpsh is running in non-interactive mode (e.g., as part of UNIX
pipe, etc).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6510 26435e26-4b25-483a-942d-477236992e84