Remove the definition and usage of PIM_HELLO_ADDRESS_LIST_PRIVATE_OPTION (defined to have value of 65001 from the private option space).
Previously PIM Hello option 65001 has been used by Cisco and pim6sd
to send the list of secondary addresses per interface.
The lastest PIM-SM I-D defines option 24 for this purpose (which we
have implemented already).

Previously we accepted the 65001 option so we could talk with pim6sd
and IPv6 Cisco, but recent versions of those implementations
already implement the spec-defined option 24.

It is not worth to continue dealing with 65001, because currently both
pim6sd and Cisco send the same secondary addresses twice (inside
option 24 and then again inside option 65001).
Dealing with duplicated secondary addresses spread across different
options is going to complicate the implementation, and anyway
in the long run we don't want to deal with options from the private
numbering space.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6518 26435e26-4b25-483a-942d-477236992e84