Initial support for filters with versions.  Basically its like the other policy filters, but the filter will frist read from the route a pointer to a filter to execute, and it will run that.
It is up to the routes / implementor of varmap, to keep these pointers reference
counted.  Once the reference count reaches 0, the filter will be destroyed.

If a route has a filter of 0, it will be assigned the latest filter.

BGP now uses these filters.

TODO: proper route pushing in BGP!


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6532 26435e26-4b25-483a-942d-477236992e84