If we are deleting a subtree, then mark the whole subtree so later the scanning of the tree will invoke the appropriate method for each node.
Note that this changes the %delete semantics for the rtrmgr template
trees (per other developers' request).

Now if a node that is deleted does not have a %delete command, then
the %delete commands of its children are called instead.
This rule is applied recursively for each child that does not have
a %delete command.

Previously, we had to add practically almost everywhere %delete methods
(in each level of the template tree hierarchy) otherwise deleting
a non-leaf node with no %delete method wouldn't invoke the %delete methods
below it.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6534 26435e26-4b25-483a-942d-477236992e84