Modify some of the logic that considers whether a module is active. Previously, if we deleted practically all child nodes of the module then the module is considered non active anymore, and would be shutdown.
For example, if we delete the last configured interface:
"delete interfaces interface foo"

then committing the deletion will shutdown module "interfaces".
This may not be the desired intention so it is better not to
shutdown the module.

Now the above command will delete only interface foo, and module
"interfaces" will continue running.
If we really want to shutdown the module, the correct command
is "delete interfaces".

Bug found by:	"Michael Larson" <mike AT lrlart.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6543 26435e26-4b25-483a-942d-477236992e84