Fix some bugs: * Propagate flags in static filters. * sync varrw in case of empty policies.
Possibly introduced a memleak in bgp_varrw.cc in the case in which routes have
only policytags / filter pointers updated.  In this case, the is_changed() flags
is not set, so the route won't be cached [as only "metarouting info" changed]
[actually, maybe we should cache modified policytags?]

Anyway, there have been discussions on re-designing the subnet route structure,
so that should come soon and "solve" this.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6545 26435e26-4b25-483a-942d-477236992e84