Diff reduction with Windows branch:
Introduce a global switch _comm_squelch_accept. If this is set to a
non-zero value, then the libcomm library will clear WinSock event
notification masks on sockets during accept and close operations.

This is done by calling WSAEventSelect() with empty arguments, which is
documented to clear such masks.

The reason for doing this is to prevent the WinSock infrastructure
from signalling event conditions before clients of EventLoop have had
a chance to wire up I/O event callbacks to service them.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6603 26435e26-4b25-483a-942d-477236992e84