1) A new event loading done. Still todo cause a Network-LSA to be generated if this is the DR.
2) When LSAs are received check them against the link state request
list. Also when queuing outgoing LSAs check them against the link
state request list. If the list becomes empty generate the event
loading done.

3) Totally change the way that router links are sent to the
area_router. Rather than adding and deleting links the whole list is
sent if a change has occured. The peer will only send a new list of
router links if a change has occurred. In the new model in the state
backup for OSPFv2 no links were generated. Previously links were left
behind. In the OSPFv3 case no links are generatated in the backup
case. Perhaps incorrectly in the OSPFv2 case backup is treated the
same as DR_other.

4) A new method "on_link_state_request_list" required by the area
router to see if an LSA is on the link state request list.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@6711 26435e26-4b25-483a-942d-477236992e84